+++ maven-plugin/src/main/java/com/aeontronix/enhancedmule/tools/PrepareExchangePublish.java
@@ -4,6 +4,7 @@
 
 package com.aeontronix.enhancedmule.tools;
 
+import com.aeontronix.enhancedmule.tools.anypoint.NotFoundException;
 import com.aeontronix.enhancedmule.tools.anypoint.Organization;
 import org.apache.maven.artifact.Artifact;
 import org.apache.maven.artifact.repository.MavenArtifactRepository;
@@ -13,16 +14,21 @@ import org.apache.maven.plugins.annotations.LifecyclePhase;
 import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
 import org.apache.maven.project.MavenProject;
+import org.slf4j.Logger;
 
+import java.io.IOException;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
+
+import static org.slf4j.LoggerFactory.getLogger;
 
 /**
  * Prepare a project for deployment to anypoint exchange maven by changing group ids to org ids
  */
 @Mojo(name = "prepare-publish", defaultPhase = LifecyclePhase.VALIDATE)
 public class PrepareExchangePublish extends AbstractOrganizationalMojo {
-    @Parameter(property = "anypoint.adddistmngmt",defaultValue = "true")
+    private static final Logger logger = getLogger(PrepareExchangePublish.class);
+    @Parameter(property = "anypoint.adddistmngmt", defaultValue = "true")
     private boolean addDistributionManagement;
     @Parameter(property = "anypoint.serverid", defaultValue = "anypoint-exchange-v2")
     private String serverId;
@@ -37,28 +43,42 @@ public class PrepareExchangePublish extends AbstractOrganizationalMojo {
 
     @Override
     protected void doExecute() throws Exception {
-        Organization organization = getOrganization();
+        Organization organization = null;
+        try {
+            organization = getOrganization();
+        } catch (NotFoundException | IOException e) {
+            logger.warn("Unable to login to exchange to retrieve org id, skipping prepare-publish");
+            return;
+        }
+        final Artifact projectArtifact = project.getArtifact();
         if (!skip) {
             String groupId = organization.getId();
+            logger.debug("Changing project groupId to " + groupId);
             project.setGroupId(groupId);
             if (project.getAttachedArtifacts() != null) {
                 for (Artifact attachedArtifact : project.getAttachedArtifacts()) {
-                    attachedArtifact.setGroupId(groupId);
+                    try {
+                        logger.debug("Changing attached artifact type {}:{} groupId to {}", attachedArtifact.getType(), attachedArtifact.getClassifier(), groupId);
+                        attachedArtifact.setGroupId(groupId);
+                    } catch (UnsupportedOperationException e) {
+                        logger.debug("Unable to change attached artifact groupId", e);
+                    }
                 }
             }
-            if (project.getArtifact() != null) {
-                project.getArtifact().setGroupId(groupId);
+            if (projectArtifact != null) {
+                logger.debug("Changing project artifact type {}:{} groupId to {}", projectArtifact.getType(), projectArtifact.getClassifier(), groupId);
+                projectArtifact.setGroupId(groupId);
             }
         }
-        if (updateVersionIfSnapshot && project.getArtifact() != null && project.getArtifact().isSnapshot()) {
-            if( snapshotVersionSuffic == null ) {
+        if (updateVersionIfSnapshot && projectArtifact != null && projectArtifact.isSnapshot()) {
+            if (snapshotVersionSuffic == null) {
                 snapshotVersionSuffic = LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyyMMddHHmmss"));
             }
             snapshotVersionSuffic = project.getVersion() + "-" + snapshotVersionSuffic;
             project.setVersion(snapshotVersionSuffic);
-            if (project.getArtifact() != null) {
-                project.getArtifact().setVersion(snapshotVersionSuffic);
-                project.getArtifact().setVersionRange(VersionRange.createFromVersion(snapshotVersionSuffic));
+            if (projectArtifact != null) {
+                projectArtifact.setVersion(snapshotVersionSuffic);
+                projectArtifact.setVersionRange(VersionRange.createFromVersion(snapshotVersionSuffic));
             }
         }
         if (project.getDistributionManagement() == null && addDistributionManagement) {